python-service/requirements.txt
flask
flask-cors
face_recognition
numpy
Pillow
opencv-python
mysql-connector-python